// mobile-app/services/enums/TruckColorsEnum.ts
export enum TruckColors {
  Amarela = "Amarela",
  Azul = "Azul",
  Bege = "Bege",
  Branca = "Branca",
  Cinza = "Cinza",
  Dourada = "Dourada",
  Grená = "Grená",
  Laranja = "Laranja",
  Marrom = "Marrom",
  Prata = "Prata",
  Preta = "Preta",
  Rosa = "Rosa",
  Roxa = "Roxa",
  Verde = "Verde",
  Vermelha = "Vermelha",
  AzulClara = "Azul Clara",
  AzulEscura = "Azul Escura",
  VerdeClara = "Verde Clara",
  VerdeEscura = "Verde Escura",
  CinzaClara = "Cinza Clara",
  CinzaEscura = "Cinza Escura",
  SemCorPredominante = "Sem Cor Predominante",
}